requirements.txt
fastapi
uvicorn[standard]
python-multipart
numpy<2
pyannote.audio==3.1.1
torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu